refactored recreation logic to be inside the sub group selector class hierarchy
